+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - Export `make_xwalk_list()` helper function
 - Reexport functions from `{esri2sf}` and `{sfext}` packages
+- Add new `cache_location_data()` function
 
 # getdata 0.1.0.9002 (2023-05-02)